Update annotations.py
Added variant_report() code to produce a JSON with VRS information for the variant in question. NOTE THAT 1) I have not been able to get with open() to work for gs:// paths and there are still some left-behinds from when I was running this locally.

There are some pretty obvious changes I still need to make AND we need to discuss what we actually want this to return. Do we want it to write out the json or return it as a variable or?

Much to discuss, but I just wanted to get the code I do have up in the correct place !